requirements.txt
beautifulsoup4==4.9.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
gTTS==2.1.1
gTTS-token==1.1.3
idna==2.9
PyAudio==0.2.11
pyserial==3.4
requests==2.23.0
six==1.14.0
soupsieve==2.0
SpeechRecognition==3.8.1
urllib3==1.25.9
